agenteScraping/requirements.txt
langchain>=0.1.0,<0.2.0
langchain-anthropic>=0.1.0,<0.2.0
anthropic>=0.17.0,<1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
python-dotenv==1.0.0